[Lens] moving store loading to middleware (#106872)
