+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/workspace_panel/chart_switch.tsx
@@ -20,7 +20,13 @@ import {
 } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n/react';
-import { Visualization, FramePublicAPI, Datasource, VisualizationType } from '../../../types';
+import {
+  Visualization,
+  FramePublicAPI,
+  VisualizationType,
+  VisualizationMap,
+  DatasourceMap,
+} from '../../../types';
 import { getSuggestions, switchToSuggestion, Suggestion } from '../suggestion_helpers';
 import { trackUiEvent } from '../../../lens_ui_telemetry';
 import { ToolbarButton } from '../../../../../../../src/plugins/kibana_react/public';
@@ -44,9 +50,9 @@ interface VisualizationSelection {
 }
 
 interface Props {
-  visualizationMap: Record<string, Visualization>;
   framePublicAPI: FramePublicAPI;
-  datasourceMap: Record<string, Datasource>;
+  visualizationMap: VisualizationMap;
+  datasourceMap: DatasourceMap;
 }
 
 type SelectableEntry = EuiSelectableOption<{ value: string }>;
@@ -55,7 +61,7 @@ function VisualizationSummary({
   visualizationMap,
   visualization,
 }: {
-  visualizationMap: Record<string, Visualization>;
+  visualizationMap: VisualizationMap;
   visualization: {
     activeId: string | null;
     state: unknown;
