+++ x-pack/plugins/lens/public/mocks.tsx
@@ -16,6 +16,7 @@ import moment from 'moment';
 import { Provider } from 'react-redux';
 import { act } from 'react-dom/test-utils';
 import { ReactExpressionRendererProps } from 'src/plugins/expressions/public';
+import { DeepPartial } from '@reduxjs/toolkit';
 import { LensPublicStart } from '.';
 import { visualizationTypes } from './xy_visualization/types';
 import { navigationPluginMock } from '../../../../src/plugins/navigation/public/mocks';
@@ -35,7 +36,7 @@ import {
 import { LensAttributeService } from './lens_attribute_service';
 import { EmbeddableStateTransfer } from '../../../../src/plugins/embeddable/public';
 
-import { makeConfigureStore, getPreloadedState, LensAppState } from './state_management/index';
+import { makeConfigureStore, LensAppState, LensState } from './state_management/index';
 import { getResolvedDateRange } from './utils';
 import { presentationUtilPluginMock } from '../../../../src/plugins/presentation_util/public/mocks';
 import { DatasourcePublicAPI, Datasource, Visualization, FramePublicAPI } from './types';
@@ -82,6 +83,11 @@ export function createMockVisualization(): jest.Mocked<Visualization> {
   };
 }
 
+const visualizationMap = {
+  vis: createMockVisualization(),
+  vis2: createMockVisualization(),
+};
+
 export type DatasourceMock = jest.Mocked<Datasource> & {
   publicAPIMock: jest.Mocked<DatasourcePublicAPI>;
 };
@@ -125,6 +131,13 @@ export function createMockDatasource(id: string): DatasourceMock {
     checkIntegrity: jest.fn((_state) => []),
   };
 }
+
+const mockDatasource: DatasourceMock = createMockDatasource('testDatasource');
+const mockDatasource2: DatasourceMock = createMockDatasource('testDatasource2');
+const datasourceMap = {
+  testDatasource2: mockDatasource2,
+  testDatasource: mockDatasource,
+};
 
 export function createExpressionRendererMock(): jest.Mock<
   React.ReactElement,
@@ -401,17 +414,21 @@ export function makeLensStore({
     data = mockDataPlugin();
   }
   const lensStore = makeConfigureStore(
-    getPreloadedState({
-      ...defaultState,
-      searchSessionId: data.search.session.start(),
-      query: data.query.queryString.getQuery(),
-      filters: data.query.filterManager.getGlobalFilters(),
-      resolvedDateRange: getResolvedDateRange(data.query.timefilter.timefilter),
-      ...preloadedState,
-    }),
     {
-      data,
-    }
+      lensServices: { ...makeDefaultServices(), data },
+      datasourceMap,
+      visualizationMap,
+    },
+    {
+      lens: {
+        ...defaultState,
+        searchSessionId: data.search.session.start(),
+        query: data.query.queryString.getQuery(),
+        filters: data.query.filterManager.getGlobalFilters(),
+        resolvedDateRange: getResolvedDateRange(data.query.timefilter.timefilter),
+        ...preloadedState,
+      },
+    } as DeepPartial<LensState>
   );
 
   const origDispatch = lensStore.dispatch;
