+++ x-pack/plugins/lens/public/state_management/types.ts
@@ -5,11 +5,15 @@
  * 2.0.
  */
 
+import { VisualizeFieldContext } from 'src/plugins/ui_actions/public';
+import { EmbeddableEditorState } from 'src/plugins/embeddable/public';
 import { Filter, Query, SavedQuery } from '../../../../../src/plugins/data/public';
 import { Document } from '../persistence';
 
 import { TableInspectorAdapter } from '../editor_frame_service/types';
 import { DateRange } from '../../common';
+import { LensAppServices } from '../app_plugin/types';
+import { DatasourceMap, VisualizationMap } from '../types';
 
 export interface PreviewState {
   visualization: {
@@ -48,4 +52,12 @@ export type DispatchSetState = (
 
 export interface LensState {
   lens: LensAppState;
+}
+
+export interface LensStoreDeps {
+  lensServices: LensAppServices;
+  datasourceMap: DatasourceMap;
+  visualizationMap: VisualizationMap;
+  initialContext?: VisualizeFieldContext;
+  embeddableEditorIncomingState?: EmbeddableEditorState;
 }
